版本更新到7.5.0

+++ GXLAMapNaviKit/Vendors/AMapNaviKit.framework/Headers/AMapNaviLocation.h
@@ -21,7 +21,7 @@
 @property (nonatomic, assign) double altitude;
 
 ///方向
-@property (nonatomic, assign) NSInteger heading;
+@property (nonatomic, assign) double heading;
 
 ///速度(km/h)
 @property (nonatomic, assign) NSInteger speed;
